exercises/src/main/java/strategy/SocialNetwork.java
package strategy;

import java.util.List;
import java.util.function.Consumer;
import java.util.function.Predicate;
import java.util.stream.Collectors;

public class SocialNetwork {

    private List<Member> members;

    public SocialNetwork(List<Member> members) {
        this.members = members;
    }

    public List<Member> findMembersBy(Predicate<Member> predicate) {
        return members.stream().filter(predicate).collect(Collectors.toList());
    }

    public void applyToSelectedMembers(Predicate<Member> predicate, Consumer<Member> consumer) {
        members.stream().filter(predicate).forEach(consumer);
    }

}
